pyproject.toml
[project]
name = "polymarket-position-notifier"
version = "0.1.0"
description = ""
requires-python = ">=3.11"
dependencies = [
    "requests",
    "pytest",
    "py-clob-client",
    "python-dotenv",
    "pymongo>=4.14.0",
    "pyyaml>=6.0.2",
    "schedule>=1.2.2",
    "tenacity>=9.1.2",
]
